+++ src/test/typopo.test.js
@@ -67,7 +67,9 @@ describe('Tests that all modules are plugged in', () => {
 		"under Law §1782": "under Law § 1782",
 		// copyright
 		"Company (c)2017": "Company © 2017",
-		
+		// spund recording copyright
+		"Company (p)2017": "Company ℗ 2017",
+
 	};
 
 	Object.keys(testCase).forEach((key) => {