+++ src/typopo.js
@@ -21,6 +21,7 @@ import {fixSymbols} from "./lib/symbols/replacements";
 import {fixMultiplicationSign} from "./lib/symbols/multiplication-sign";
 import {fixSectionSign} from "./lib/symbols/section-sign";
 import {fixCopyright} from "./lib/symbols/copyright";
+import {fixSoundRecordingCopyright} from "./lib/symbols/sound-recording-copyright";
 import {fixAbbreviations} from "./lib/words/abbreviations";
 import {fixCase} from "./lib/words/case";
 import {excludeExceptions,
@@ -66,6 +67,7 @@ export function fixTypos(string, locale, configuration) {
 	string = fixMultiplicationSign(string, currentLocale);
 	string = fixSectionSign(string, currentLocale);
 	string = fixCopyright(string, currentLocale);
+	string = fixSoundRecordingCopyright(string, currentLocale);
 
 	// words
 	string = fixCase(string, currentLocale);